🧹 Remove old components, structure new components
