+++ src/assets/js/main.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Comments from "../../components/Comments";
+import Comments from "../../components/Comment/Comments";
 import "../css/index.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
