+++ src/components/User/Avatar.jsx
@@ -1,7 +1,9 @@
-export const Avatar = ({ user, className }) => (
+const Avatar = ({ user, className }) => (
   <img
-    src={user.image}
+    src={user.image.png}
     alt={`${user.username}'s Avatar`}
     className={`h-8 w-8 rounded-full shadow-inner shadow-gray ${className}`}
   />
 );
+
+export default Avatar;
